KOGITO-2843: Make ProcessDescriptor generic - Component rename
