+++ ui-packages/packages/management-console/src/components/Organisms/ProcessDetails/ProcessDetails.tsx
@@ -14,7 +14,7 @@ import {
 import React from 'react';
 import { LevelDownAltIcon, LevelUpAltIcon } from '@patternfly/react-icons';
 import { Link } from 'react-router-dom';
-import { ProcessDescriptor, GraphQL, EndpointLink } from '@kogito-apps/common';
+import { ItemDescriptor, GraphQL, EndpointLink } from '@kogito-apps/common';
 import { stateIconCreator } from '../../../utils/Utils';
 import ProcessInstance = GraphQL.ProcessInstance;
 
@@ -132,7 +132,7 @@ const ProcessDetails: React.FC<IOwnProps> = ({ data, from }) => {
                     content={data.ProcessInstances[0].parentProcessInstance.id}
                   >
                     <Button variant="link" icon={<LevelUpAltIcon />}>
-                      <ProcessDescriptor
+                      <ItemDescriptor
                         processInstanceData={
                           data.ProcessInstances[0].parentProcessInstance
                         }
@@ -154,7 +154,7 @@ const ProcessDetails: React.FC<IOwnProps> = ({ data, from }) => {
                     >
                       <Tooltip content={child.id}>
                         <Button variant="link" icon={<LevelDownAltIcon />}>
-                          <ProcessDescriptor processInstanceData={child} />
+                          <ItemDescriptor processInstanceData={child} />
                         </Button>
                       </Tooltip>
                     </Link>
