+++ ui-packages/packages/management-console/src/components/Templates/ProcessDetailsPage/tests/ProcessDetailsPage.test.tsx
@@ -51,7 +51,7 @@ jest.mock('@patternfly/react-icons', () => ({
 
 jest.mock('@kogito-apps/common', () => ({
   ...jest.requireActual('@kogito-apps/common'),
-  ProcessDescriptor: () => {
+  ItemDescriptor: () => {
     return <MockedComponent />;
   },
   KogitoSpinner: () => {
